+++ modules/ui_extra_networks.py
@@ -300,8 +300,9 @@ class ExtraNetworksPage:
         self.html = ''
         self.create_items(tabname)
         versions = sorted({item.get("version", "") for item in self.items if item.get("version")})
-        if 'ref' in versions:
-            versions.remove('ref')
+        for v in ['ref', 'reference', 'ready', 'download']:
+            if v in versions:
+                versions.remove(v)
         versions_html = ''
         for ver in versions:
             versions_html += f'<button class="lg secondary gradio-button custom-button network-model" onclick="extraNetworksFilterVersion(event)">{html.escape(ver)}</button><br>'